requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.12
langgraph==0.0.20
chromadb==0.4.22
sentence-transformers==2.2.2
numpy==1.24.3
torch==2.1.2
transformers==4.36.2
tiktoken==0.5.2
ollama==0.1.7
PyYAML==6.0.1
python-dotenv==1.0.1
requests==2.31.0 